kinda working adding a new account need some extra functions tho

+++ src/main/java/io/swagger/repository/AccountRepository.java
@@ -4,6 +4,6 @@ import io.swagger.model.Account;
 import io.swagger.model.User;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AccountRepository extends CrudRepository<Account, Long>{
+public interface AccountRepository extends CrudRepository<Account, Integer>{
 
 }
